pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "mlops-template"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "scikit-learn>=1.5.0",
    "xgboost>=2.0.3",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "mlflow>=2.14.0",
    "evidently>=0.4.30",
    "prometheus-client>=0.20.0",
    "apache-airflow>=2.9.2",
    "great-expectations>=0.18.0",
    "sqlalchemy>=2.0.0",
    "boto3>=1.34.0",
    "structlog>=24.1.0",
    "pyarrow>=16.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "hypothesis>=6.100.0",
    "pytest-cov>=5.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
